+++ MoreShipUpgrades/UpgradeComponents/TierUpgrades/Vehicle/ImprovedSteering.cs
@@ -3,12 +3,9 @@ using MoreShipUpgrades.Misc.TerminalNodes;
 using MoreShipUpgrades.Misc;
 using MoreShipUpgrades.Misc.Upgrades;
 using MoreShipUpgrades.Misc.Util;
-using System;
-using System.Collections.Generic;
-using System.Text;
 using UnityEngine;
 
-namespace MoreShipUpgrades.UpgradeComponents.TierUpgrades.AttributeUpgrades
+namespace MoreShipUpgrades.UpgradeComponents.TierUpgrades.Vehicle
 {
     internal class ImprovedSteering : TierUpgrade
     {
@@ -70,7 +67,7 @@ namespace MoreShipUpgrades.UpgradeComponents.TierUpgrades.AttributeUpgrades
         }
         public override string GetDisplayInfo(int initialPrice = -1, int maxLevels = -1, int[] incrementalPrices = null)
         {
-            System.Func<int, float> infoFunction = level => UpgradeBus.Instance.PluginConfiguration.IMPROVED_STEERING_TURNING_SPEED_INITIAL_INCREASE.Value + (level * UpgradeBus.Instance.PluginConfiguration.IMPROVED_STEERING_TURNING_SPEED_INCREMENTAL_INCREASE.Value);
+            System.Func<int, float> infoFunction = level => UpgradeBus.Instance.PluginConfiguration.IMPROVED_STEERING_TURNING_SPEED_INITIAL_INCREASE.Value + level * UpgradeBus.Instance.PluginConfiguration.IMPROVED_STEERING_TURNING_SPEED_INCREMENTAL_INCREASE.Value;
             const string infoFormat = "LVL {0} - ${1} - Company Cruiser vehicle's steering speed (turning speed) is increased by {2}.\n";
             return Tools.GenerateInfoForUpgrade(infoFormat, initialPrice, incrementalPrices, infoFunction);
         }