+++ MoreShipUpgrades/UpgradeComponents/TierUpgrades/Vehicle/VehiclePlating.cs
@@ -5,7 +5,7 @@ using MoreShipUpgrades.Misc.Upgrades;
 using MoreShipUpgrades.Misc.Util;
 using UnityEngine;
 
-namespace MoreShipUpgrades.UpgradeComponents.TierUpgrades.AttributeUpgrades
+namespace MoreShipUpgrades.UpgradeComponents.TierUpgrades.Vehicle
 {
     internal class VehiclePlating : TierUpgrade
     {
@@ -69,7 +69,7 @@ namespace MoreShipUpgrades.UpgradeComponents.TierUpgrades.AttributeUpgrades
         }
         public override string GetDisplayInfo(int initialPrice = -1, int maxLevels = -1, int[] incrementalPrices = null)
         {
-            System.Func<int, float> infoFunction = level => UpgradeBus.Instance.PluginConfiguration.VEHICLE_PLATING_HEALTH_INITIAL_INCREASE.Value + (level * UpgradeBus.Instance.PluginConfiguration.VEHICLE_PLATING_HEALTH_INCREMENTAL_INCREASE.Value);
+            System.Func<int, float> infoFunction = level => UpgradeBus.Instance.PluginConfiguration.VEHICLE_PLATING_HEALTH_INITIAL_INCREASE.Value + level * UpgradeBus.Instance.PluginConfiguration.VEHICLE_PLATING_HEALTH_INCREMENTAL_INCREASE.Value;
             const string infoFormat = "LVL {0} - ${1} - Company Cruiser vehicle's maximum health is increased by {2}.\n";
             return Tools.GenerateInfoForUpgrade(infoFormat, initialPrice, incrementalPrices, infoFunction);
         }