+++ ChangeLog.md
@@ -5,6 +5,8 @@
   - You can configure the amount of upgrades that can be picked to show up.
   - You can configure to show the upgrades already purchased.
   - You can configure when should the upgrades be randomized again (per quota, per moon routing or per moon landing).
+- Implemented "Fluffy Seats" tier upgrade which adds player damage mitigation when driving the Company Cruiser vehicle.
+- Implemented "Ignition Coil" tier upgrade which increases the chance of igniting on the Company Cruiser vehicle.
 - Fixed issue with some relevant components not being saved correctly, leading to constantly changing per save boot.
 
 </details>